refactor(agent): extract context-graph lifecycle into mixin holder
First real mixin-holder migration. Moves the context-graph lifecycle
cluster (createContextGraph, registerContextGraph, inviteToContextGraph,
inviteAgentToContextGraph, removeAgentFromContextGraph, renameContextGraph,
getContextGraphAllowedAgents — ~1.5k lines) out of dkg-agent.ts into
ContextGraphMethods (extends DKGAgentBase), assembled onto DKGAgent via
applyMixins + `interface DKGAgent extends ContextGraphMethods`.

Methods take `this: DKGAgent` so cross-calls resolve against the composed
class. 21 sibling methods the cluster calls were widened private->public
(they become public anyway as their own clusters migrate). Pure relocation.

dkg-agent.ts 21400 -> 19884. Typecheck clean; unit (123) + agent.test.ts
(126) + CG suites (e2e/allowlist/enumerate/discovery, 80) green.
